M.I.A pre-alpha 0.2.0
It has the Logarithmic Random Ponderated Selector's algorithm. This will be very important to whole the system. The algorithm takes a list of elements and selects randomly one, but this selection is ponderated using logarithms; the first item has more probability to be selected than the second, and the second more than the third, etc. It will be important in selections when we want to give more frecuency to some events. It shows also a graphic with the curve form to demonstrate the logarithmic curve influence.