Grammar Components v.1.0.0
Free-Context Grammar added, capable to generate sentences of FOUR types: 1. Simple Declarative Sentences, 2. Simple Interrogative Sentences, 3. Exclamative Sentences, 4. Imperative Sentences. THIS IS A FIRST APROXIMATION, It takes from csv data of PartOfSpeech dictionaries the words to replace into the generative grammars. Some sentences seems to have no sense because the Production rules. It could be improved in kind of Verbs, or maybe just sintactical structures. Also Sustantives and Verbs are a little bit, it could be great to add a really big dictionaries for it