MIA Components v.1.4.0
Function degree_of_chord added to compute the degree of a chord on the context of a key/scale. It will be necessary to the implementtation of the Genetic Algorithm for Progressions. This function gives a string with the degree in roman notation adding accidents as b, # or special things for complex chords