MIA Algorithms b.1.0.0
Genetic Algorithm to generate Chord Progressions [IN PROGRESS]. It has the full logic of work, including most important the initialization method. It generates the first population using LWRS (Logarithmic Weighted Random Selector) Algorithm for selecting the preference in chords in the key or out the key, and implements a new selection method called Complexity Weighted Random Selector (CWRS) that decides if it will apply a complexity to simple major/minor (and also dim and aug) chords. After the step of selecting the chords to initializate a progression, the CWRS algorithm takes a parameter [complexity] between 0 and 100, based on; if it is 0, the probability to take complex chords as 7th, maj7, 9th, 11th, 13th, 6th, etc. is 0. If complexity is 100, the probability to take this complex chords are the same for the rest of the just major/minor chords

TODO: APPLY TE RULES FOR VALIDATE AN EVALUATE A CHORD PROGRESSION INTO THE CODE OF OBJECTIVE FUNCTION, BASED ON THE RULES PLASMED ON THE THESIS DOCUMENT